+++ src/ts/typings/window.d.ts
@@ -1,5 +1,4 @@
-import { type Select2Plugin } from 'select2';
-
+import type { StancerSettings } from "./stancer";
 declare global {
   interface AdminData{
     confirmMessage: string;
@@ -13,41 +12,131 @@ declare global {
     message?: string | null;
   }
 
-  interface StancerData {
-    changePaymentMethod?: ChangePaymentMethod
-    checkout_url: string;
-    initiate: string;
+  const stancer_data: StancerData;
+  const wc_checkout_params: WooCommerceCheckoutParams;
+
+  type CheckoutResponse = CheckoutResponseFailure | CheckoutResponseSuccess
+  type PaymentMethods = { stancer: StancerSettings }
+
+  interface ChangePaymentData {
+    nonce?: string;
+    subscription: string;
+    action: 'information' | 'initiate' | 'validate';
   }
   interface ChangePaymentMethod {
     nonce: string;
-    url:string;
-  }
-  interface WooCommerceCheckoutParams {
-    checkout_url: string;
+    url: string;
   }
 
-  interface JQuery<TElement = HTMLElement> {
-    block: (params?: BlockParams) => this;
-    selectWoo: Select2Plugin<TElement>;
-    unblock: () => this;
+  interface CheckoutResponseBase {
+    card?: string;
+    messages?: string;
+    refresh: boolean;
+    reload: boolean;
   }
 
-  interface JQueryStatic {
-    scroll_to_notices: (element: JQuery) => this;
+  interface CheckoutResponseFailure extends CheckoutResponseBase {
+    result: 'failure';
+    reason: string
   }
 
-  interface Window {
-    stancer_admin: AdminData;
-    stancer: StancerData;
-    wc_checkout_params: WooCommerceCheckoutParams;
+  interface CheckoutResponseReact {
+    redirect: string
+    result: string
+    receipt: string
   }
 
-  const stancer: StancerData;
-  const wc_checkout_params: WooCommerceCheckoutParams;
+  interface CheckoutResponseSuccess extends CheckoutResponseBase {
+    order_id: number | string;
+    receipt: string;
+    redirect: string;
+    result: 'success';
+  }
+
+  interface ListenerData extends ListenerDataBase {
+    button: string;
+  }
+
+  interface ListenerDataBase {
+    route: {
+      url?: string;
+      path: string;
+    }
+    data: ChangePaymentData | (() => string) | string;
+  }
+
+  interface MessageData {
+    height: number;
+    status: 'error' | 'finished' | 'init' | 'secure-auth-end' | 'secure-auth-error' | 'secure-auth-start';
+    url: string & Location;
+    width: number;
+  }
+
+  interface React { }
 
   interface Redirection {
     receipt: string;
   }
+
+  interface ServerCallData extends ListenerDataBase {
+    responseCallBack: (arg0: CheckoutResponse) => void;
+  }
+
+  interface StancerData {
+    changePaymentMethod?: ChangePaymentMethod
+    initiate: string;
+  }
+
+  interface stancerWindows {
+    stancer_iframe: (result: CheckoutResponseSuccess | CheckoutResponseReact) => void;
+    stancer_onSubmit: (arg0: ListenerData, arg1?: () => boolean) => void;
+    stancer_changePaymentMethodOrCallback: (arg0: () => void) => () => void;
+    stancer_callServer: ({ data, route, responseCallBack }: ServerCallData) => void;
+    stancer_paymentMethodHasBeenChanged: (arg0: MessageData) => boolean;
+  }
+
+  interface WCSettings {
+    currency: {
+      code: string;
+    }
+    paymentMethodData: PaymentMethods;
+    description: string;
+  }
+
+  interface WordPressEnvironment {
+    htmlEntities: {
+      decodeEntities: (arg0: string) => string
+    }
+    apiFetch(arg0:{
+      path: string;
+      method:string;
+      data: ChangePaymentData | (() => string) | string;
+    }):Promise<CheckoutResponse>;
+    apiFetch(arg0: {
+      path: '/wc/store/v1/checkout',
+      method: 'POST',
+      data: CompletePaymentData,
+    }):Promise<BlockApiResponse>;
+  }
+
+  interface WooCommerceCheckoutParams {
+    checkout_url: string;
+  }
+
+  interface WooCommerceEnvironment {
+    wcBlocksRegistry: {
+      registerPaymentMethod: (arg0: object) => void
+    }
+    wcSettings: WCSettings
+  }
+  interface Window extends stancerWindows {
+    stancer_admin: AdminData;
+    stancer: StancerData;
+    wc : WooCommerceEnvironment;
+    wc_checkout_params: WooCommerceCheckoutParams;
+    wp: WordPressEnvironment;
+    wcSettings: WCSettings;
+  }
 }
 
-export {};
+export { };