+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+- Support of woocommerce checkout blocks. (CMS-68)
+
 ### Changed
 - Change the cb logo to the new logo (CMS-173)
+
 
 ## [1.2.6] - 2024-08-08